moving useVotingStatus to api module
