+++ src/screen/ReferendaScreen.tsx
@@ -8,7 +8,7 @@ import * as React from 'react';
 import {FlatList, StyleSheet} from 'react-native';
 import {useBlockTime} from 'src/api/hooks/useBlockTime';
 import {useReferendums} from 'src/api/hooks/useReferendums';
-import {useBestNumber} from 'src/hook/useVotingStatus';
+import {useBestNumber} from 'src/api/hooks/useVotingStatus';
 import globalStyles, {standardPadding} from 'src/styles';
 
 export function ReferendaScreen() {
